Make FileStore thread-safe Enables running the soundtrack extraction and combining in parallel, which is also included in this commit. Move oggCat to application's root directory (for easier calls to generic OS libraries).
It's been running for 10 minutes now without error, so I think I can commit it safely now.

+++ RuneScapeCacheTools/Cache/RuneTek5/Index.cs
@@ -19,8 +19,8 @@ namespace Villermen.RuneScapeCacheTools.Cache.RuneTek5
 		public Index(byte[] data)
 		{
 			var reader = new BinaryReader(new MemoryStream(data));
-			Size = (int) reader.ReadUInt24BigEndian();
-			Sector = (int) reader.ReadUInt24BigEndian();
+			Size = reader.ReadUInt24BigEndian();
+			Sector = reader.ReadUInt24BigEndian();
 		}
 
 		/// <summary>
